+++ src/_modules/footer/footer.js
@@ -1,8 +1,0 @@
-'use strict';
-
-export default class Footer {
-  constructor() {
-    this.name = 'footer';
-    console.log('%s footer', this.name);
-  }
-}