+++ src/_scripts/main.js
@@ -1,8 +1,1 @@
-// Main javascript entry point
-// Should handle bootstrapping/starting application
-
-'use strict';
-
-$(() => {
-  console.log('Welcome to Yeogurt!');
-});
+console.log('welcome wanglang');